Moved AssignedByValParameterInspectionTests for QuickFixes to respective files
